OneDrive/Desktop/sociologie/learning_platform/requirements_dev.txt
Django==4.2.8
djangorestframework==3.14.0
python-decouple==3.8
django-cors-headers==4.3.1
django-filter==23.5
django-extensions==3.2.3
pytest==7.4.3
pytest-django==4.7.0
openai>=1.0.0
requests==2.31.0
djangorestframework-simplejwt>=5.2.0
Pillow==10.1.0
